ban users from moderate cards
