+++ src/core/client/admin/components/ModerateCard/ModerateCard.tsx
@@ -1,5 +1,7 @@
 import cn from "classnames";
+import { ModerateCardContainer_comment } from "coral-admin/__generated__/ModerateCardContainer_comment.graphql";
 import { HOTKEYS } from "coral-admin/constants";
+import { GQLUSER_STATUS } from "coral-framework/schema";
 import { Localized } from "fluent-react/compat";
 import key from "keymaster";
 import React, {
@@ -15,6 +17,7 @@ import {
   Card,
   Flex,
   HorizontalGutter,
+  Tag,
   TextLink,
   Timestamp,
   Typography,
@@ -39,7 +42,7 @@ interface Props {
     id: string;
     username: string | null;
   } | null;
-  comment: PropTypesOf<typeof MarkersContainer>["comment"];
+  comment: ModerateCardContainer_comment;
   settings: PropTypesOf<typeof MarkersContainer>["settings"];
   status: "approved" | "rejected" | "undecided";
   featured: boolean;
@@ -116,7 +119,9 @@ const ModerateCard: FunctionComponent<Props> = ({
       }
       key(HOTKEYS.APPROVE, id, onApprove);
       key(HOTKEYS.REJECT, id, onReject);
-      key(HOTKEYS.BAN, id, onBan);
+      if (onBan) {
+        key(HOTKEYS.BAN, id, onBan);
+      }
       key.setScope(id);
       return () => {
         key.deleteScope(id);
@@ -125,7 +130,7 @@ const ModerateCard: FunctionComponent<Props> = ({
       key.deleteScope(id);
     }
     return () => null;
-  }, [selected]);
+  }, [selected, comment]);
 
   useEffect(() => {
     if (selected && div && div.current) {
